medicine_ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
langgraph>=1.0.0
langgraph-checkpoint-sqlite==3.0.3
streamlit
requests
pandas
python-dotenv==1.0.1

